Move to internal/common (address comment)
