--- pkg/ttlmap/go.mod
@@ -1,5 +1,0 @@
-module github.com/DataDog/opentelemetry-collector-contrib/pkg/ttlmap
-
-go 1.14
-
-require github.com/stretchr/testify v1.6.1
